# Conflicts:
#	calculator/src/main/java/com/market/banica/calculator/exception/GlobalExceptionHandler.java
#	calculator/src/main/java/com/market/banica/calculator/service/JMXServiceImpl.java
#	calculator/src/main/java/com/market/banica/calculator/service/ProductServiceImpl.java
#	calculator/src/test/java/com/market/banica/calculator/unitTests/serviceTests/ProductServiceImplTest.java
#	order-book/src/test/java/com/market/banica/order/book/service/grpc/componentTests/OrderBookComponentIT.java

+++ calculator/src/main/java/com/market/banica/calculator/service/grpc/AuroraClientSideService.java
@@ -54,7 +54,7 @@ public class AuroraClientSideService {
         }
     }
 
-    public ItemOrderBookResponse getIngredient(String productName, String clientId, int quantity) {
+    public ItemOrderBookResponse getIngredient(String productName, String clientId, long quantity) {
         LOGGER.debug("Inside getIngredient method with parameter product name - {} and client id - {}"
                 , productName, clientId);
 
@@ -100,5 +100,4 @@ public class AuroraClientSideService {
                 .setClientId(CLIENT_ID)
                 .build();
     }
-
 }
